Hold a mutex while refreshing the token
When using the default service account from the internal metadata
service there is no point in asking it for multiple tokens at the same
time.  With a high enough request rate this results in a thundering
herd of refresh requests and eventually the metadata service rejecting
the requests with HTTP 429 status code.