Move refersh locking to AuthenticationManager
This makes it central in once place.  The downside is that it is a
global lock also for the CustomServiceAccount case which could
theorethically have a lock per scope combination requested.

Smarter approaches are possible, but this is a good first approach to
avoid a thundering herd on the refresh.