Use `gcloud.cmd` to invoke `gcloud` on Windows
On Windows, the name of the `gcloud` command is `gcloud.cmd`,
while on MacOS and Linux the name is just `gcloud`.

While on the Windows Command prompt invoking `gcloud` works, this
is not the case when a process is spawned with Rust's `Command`,
causing the application to fail to execute the command.

As a solution, we introduce an OS specific variable reflecting the
actual command name for each OS.